requirements.txt
Flask~=3.1.2
Werkzeug~=3.1.4
python-dotenv~=1.2.1